loop: only auto-game_over() when there are no pieces left on deck
Choosing a piece that can't be placed shouldn't automatically end
the game, since other pieces on deck might be able to clear a row or
column that this piece can then fit in.

Also reset the game.{x,y} when move_to_clear() fails so the piece
isn't hanging off the edge of the board.